+++ TopNews/TopNewsiOS/NewsUI/Composers/MainQueueDispatchDecorator.swift
@@ -1,0 +1,44 @@
+//
+//  MainQueueDispatchDecorator.swift
+//  TopNewsiOS
+//
+//  Created by Ihwan on 19/04/22.
+//
+
+import Foundation
+import TopNews
+
+final class MainQueueDispatchDecorator<T> {
+    private let decoratee: T
+    
+    init(decoratee: T) {
+        self.decoratee = decoratee
+    }
+    
+    func dispatch(completion: @escaping () -> Void) {
+        guard Thread.isMainThread else {
+            return DispatchQueue.main.async(execute: completion)
+        }
+        
+        completion()
+    }
+}
+
+extension MainQueueDispatchDecorator: NewsLoader where T == NewsLoader {
+    func load(completion: @escaping (NewsLoader.Result) -> Void) {
+        decoratee.load { [weak self] result in
+            self?.dispatch { completion(result) }
+        }
+    }
+}
+
+extension MainQueueDispatchDecorator: NewsImageDataLoader where T == NewsImageDataLoader {
+    func loadImageData(from url: URL, completion: @escaping (NewsImageDataLoader.Result) -> Void) -> NewsImageDataLoaderTask {
+        decoratee.loadImageData(from: url) { [weak self] result in
+            self?.dispatch {
+                completion(result)
+            }
+        }
+    }
+    
+}
